put into component
